+++ libyafuse2/include/libfs/libfs.h
@@ -34,8 +34,8 @@
  */
 struct libfs_timespec
 {
-  int32_t tv_sec;
-  int32_t tv_nsec;
+  int64_t tv_sec;
+  int64_t tv_nsec;
 };
 
 struct fs_fsid_t {
@@ -43,49 +43,49 @@ struct fs_fsid_t {
 };
 
 struct fs_kstatfs {
-  int32_t f_type;
-  int32_t f_bsize;
-  uint32_t f_blocks;
-  uint32_t f_bfree;
-  uint32_t f_bavail;
-  uint32_t f_files;
-  uint32_t f_ffree;
+  int64_t f_type;
+  int64_t f_bsize;
+  uint64_t f_blocks;
+  uint64_t f_bfree;
+  uint64_t f_bavail;
+  uint64_t f_files;
+  uint64_t f_ffree;
   struct fs_fsid_t f_fsid;
-  int32_t f_namelen;
-  int32_t f_frsize;
-  int32_t f_flags;
-  int32_t f_spare[4];
+  int64_t f_namelen;
+  int64_t f_frsize;
+  int64_t f_flags;
+  int64_t f_spare[4];
 };
 
 struct fs_dirent {
-  uint32_t d_ino;
-  uint32_t d_off;
+  uint64_t d_ino;
+  int64_t d_off;
   uint16_t d_reclen;
   uint8_t d_type;
   char d_name[256];
 };
 
 struct fs_kstat {
-  uint32_t ino;
+  uint64_t ino;
   uint16_t mode;
   uint32_t nlink;
   uint32_t uid;
   uint32_t gid;
-  int32_t size;
+  int64_t size;
   struct libfs_timespec atime;
   struct libfs_timespec mtime;
   struct libfs_timespec ctime;
-  uint32_t blksize;
-  uint32_t blocks;
+  uint64_t blksize;
+  uint64_t blocks;
 };
 
 struct fs_opt_t {
-  int32_t (*mount) (const char *dev_name, const char *dir_name, const char *type, uint32_t flags, void *data);
+  int32_t (*mount) (const char *dev_name, const char *dir_name, const char *type, uint64_t flags, void *data);
   int32_t (*umount) (const char *name, int32_t flags);
   int32_t (*statfs) (const char *pathname, struct fs_kstatfs *buf);
   int32_t (*stat) (const char *filename, struct fs_kstat *statbuf);
   int32_t (*getdents) (uint32_t fd, struct fs_dirent *dirent, uint32_t count);
-  int32_t (*getcwd) (char *buf, uint32_t size);
+  int32_t (*getcwd) (char *buf, uint64_t size);
   int32_t (*chdir) (const char *filename);
 };
